add last screen create function
