+++ app/src/main/java/com/innoprog/android/feature/projects/create/presentation/ui/adapter/InputViewHolder.kt
@@ -1,0 +1,16 @@
+package com.innoprog.android.feature.projects.create.presentation.ui.adapter
+
+import androidx.recyclerview.widget.RecyclerView
+import com.innoprog.android.databinding.ItemInputViewBinding
+
+class InputViewHolder(val binding: ItemInputViewBinding) :
+    RecyclerView.ViewHolder(binding.root) {
+
+    fun bind(label: String?, hint: String, data: String?) {
+        if (label.isNullOrEmpty()) {
+            binding.itemTitle.text = label
+        }
+        binding.inputIv.setHint(hint)
+        binding.inputIv.setText(data ?: "")
+    }
+}
