+++ app/src/main/res/layout/fragment_project_create.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background_primary"
-    tools:context=".feature.projects.create.presentation.ui.CreateProjectFragment">
+    tools:context=".feature.projects.create.presentation.ui.FillAboutProjectFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/left_border"
