+++ app/src/main/res/layout/item_input_view.xml
@@ -6,7 +6,7 @@
     android:orientation="vertical">
 
     <TextView
-        android:id="@+id/financing_stage"
+        android:id="@+id/item_title"
         style="@style/TextSubtitle_1"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -15,7 +15,7 @@
         android:textColor="@color/text_primary" />
 
     <com.innoprog.android.uikit.InnoProgInputView
-        android:id="@+id/investments_description_iv"
+        android:id="@+id/input_iv"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="@dimen/margin_12"
